added predictor.py
