+++ src/drug_toxicity/__init__.py
@@ -1,70 +1,15 @@
-from __future__ import annotations
+from rdkit import RDLogger
 
-from dataclasses import dataclass, field
-from typing import TYPE_CHECKING
-
-from rdkit import Chem
+RDLogger.DisableLog('rdApp.*')
 
 from . import liver, reactions
+from .predictor import Predictor, predict_products
 from .validator import Validator
-
-if TYPE_CHECKING:
-    from collections.abc import Iterable
-
-    from rdkit.Chem.rdchem import Mol
 
 __all__ = [
     "liver",
     "reactions",
     "Predictor",
     "Validator",
+    "predict_products",
 ]
-
-
-@dataclass
-class Predictor:
-    validator: Validator = field(default_factory=Validator)
-
-    @staticmethod
-    def _products_to_smiles(products: list[Mol]) -> list[str]:
-        return [Chem.MolToSmiles(x) for x in products]
-
-    def _filter_products(self, products: list[Mol]) -> list[str]:
-        return [
-            smiles
-            for smiles in self._products_to_smiles(products)
-            if self.validator.validate(smiles)
-        ]
-
-    def predict_phase_i(self, compound: str) -> list[str]:
-        """Return the products of all phase I reactions from a compound in SMILES format."""
-        products = []
-        comp = Chem.MolFromSmiles(compound)
-        for reaction in liver.get_phase_i_reactions():
-            product = reaction(comp)
-            products.extend(product)
-
-        return self._filter_products(products)
-
-    def predict_phase_ii(self, filtered_phase_i_products: Iterable[str]) -> list[str]:
-        """Return the products of all phase II reactions from a compound in SMILES format."""
-        products = []
-        for reactant in filtered_phase_i_products:
-            reactant2 = Chem.MolFromSmiles(reactant)
-            for reaction in liver.get_phase_ii_reactions():
-                product = reaction(reactant2)
-                products.extend(product)
-
-        return self._filter_products(products) 
-
-
-def predict_products(compound_smiles: str) -> list[str]:
-    """Return the products of all phase I and phase II reactions.
-    
-    Input a compound in SMILES format. Phase I reactions are enacted on this compound,
-    followed by phase II reactions on the phase I products. 
-    A list of all products is returned.
-    """
-    predictor = Predictor()
-    phase_1_mets = predictor.predict_phase_i(compound_smiles)
-    return phase_1_mets + predictor.predict_phase_ii(phase_1_mets)
